Reorganize test files.
